# emf-control-plane/app/src/main/resources/openapi/control-plane-api.yaml
openapi: 3.0.3
info:
  title: EMF Control Plane API
  description: |
    Control Plane API for managing collections, fields, authorization, UI configuration, and packages.
    This specification serves as the single source of truth for the API contract between the control plane and UI/SDK.
  version: 1.0.0
  contact:
    name: EMF Platform Team
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.emf.example.com
    description: Production server
...
paths:
...
  /control/oidc/providers:
    get:
      tags: [OIDC]
      summary: List OIDC providers
      operationId: listOidcProviders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved OIDC providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OidcProviderDto'

    post:
      tags: [OIDC]
      summary: Add OIDC provider
      operationId: addOidcProvider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOidcProviderRequest'
      responses:
        '201':
          description: OIDC provider created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OidcProviderDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    OidcProviderDto:
      type: object
      required:
        - id
        - name
        - issuer
        - clientId
        - enabled
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        issuer:
          type: string
          format: uri
        clientId:
          type: string
        clientSecret:
          type: string
        scopes:
          type: string
          description: JSON array of scopes
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    AddOidcProviderRequest:
      type: object
      required:
        - name
        - issuer
        - clientId
      properties:
        name:
          type: string
        issuer:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        scopes:
          type: array
          items:
            type: string
        enabled:
          type: boolean
          default: true
...